Create Supp_3.R
DESCRIPTION: This code was created to generate bar plots in Figure 3 of Acevedo et al. (in review), demonstrating the percentage of angle groupings of microtubules across all selected samples and evaluating the overall CMT distribution of inner and outer CMT Angles of the Hypocotyl, using a violin and boxplot + SE bars.